Don't use usize for view counts

+++ src/layout/linear/layout_element.rs
@@ -9,7 +9,7 @@ use embedded_graphics::primitives::Rectangle;
 
 pub trait LayoutElement<LD: Orientation>: ViewChainElement {
     fn measure(&self) -> Size;
-    fn arrange(&mut self, bounds: Rectangle, orientation: &LD, count: usize) -> Rectangle;
+    fn arrange(&mut self, bounds: Rectangle, orientation: &LD, count: u32) -> Rectangle;
 }
 
 impl<V, VCE, LD> LayoutElement<LD> for Link<V, VCE>
@@ -28,7 +28,7 @@ where
         }
     }
 
-    fn arrange(&mut self, bounds: Rectangle, orientation: &LD, count: usize) -> Rectangle {
+    fn arrange(&mut self, bounds: Rectangle, orientation: &LD, count: u32) -> Rectangle {
         if VCE::IS_TERMINATOR {
             orientation.place_first(&mut self.object, bounds, count);
         } else {
@@ -50,7 +50,7 @@ impl<LD: Orientation> LayoutElement<LD> for Guard {
         Size::new(0, 0)
     }
 
-    fn arrange(&mut self, _bounds: Rectangle, _orientation: &LD, _count: usize) -> Rectangle {
+    fn arrange(&mut self, _bounds: Rectangle, _orientation: &LD, _count: u32) -> Rectangle {
         // Nothing to do
         Rectangle::new(Point::zero(), Point::zero())
     }
